requirements.txt
# install flask
flask~=2.3.2
# install mysql-connector
mysql-connector

requests~=2.31.0